# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIQE Studio API",
    "description": "Enhance, edit, organise and automate images, and train the models that do it.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/ai/plan": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "What a run would produce, and how it would run",
        "operationId": "plan_api_ai_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiRunIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiPlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/runs": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Run a model on an image; the result becomes a new image",
        "operationId": "start_run_api_ai_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiRunIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiRunStartOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AiPlanOut": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "task": {
            "type": "string",
            "enum": [
              "upscale",
              "denoise",
              "background",
              "face"
            ],
            "title": "Task"
          },
          "scale": {
            "type": "integer",
            "title": "Scale"
          },
          "device": {
            "type": "string",
            "enum": [
              "cuda",
              "cpu"
            ],
            "title": "Device"
          },
          "device_name": {
            "type": "string",
            "title": "Device Name"
          },
          "input_width": {
            "type": "integer",
            "title": "Input Width"
          },
          "input_height": {
            "type": "integer",
            "title": "Input Height"
          },
          "output_width": {
            "type": "integer",
            "title": "Output Width"
          },
          "output_height": {
            "type": "integer",
            "title": "Output Height"
          },
          "output_megapixels": {
            "type": "number",
            "title": "Output Megapixels"
          },
          "bit_depth": {
            "type": "integer",
            "title": "Bit Depth"
          },
          "has_alpha": {
            "type": "boolean",
            "title": "Has Alpha"
          },
          "tile": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tile"
          },
          "batch": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch"
          },
          "tiles": {
            "type": "integer",
            "title": "Tiles"
          },
          "calibrated": {
            "type": "boolean",
            "title": "Calibrated",
            "description": "True when the tile size comes from memory measured on this GPU."
          },
          "estimated_memory_bytes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Memory Bytes"
          },
          "disk_bytes": {
            "type": "integer",
            "title": "Disk Bytes"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "task",
          "scale",
          "device",
          "device_name",
          "input_width",
          "input_height",
          "output_width",
          "output_height",
          "output_megapixels",
          "bit_depth",
          "has_alpha",
          "tile",
          "batch",
          "tiles",
          "calibrated",
          "estimated_memory_bytes",
          "disk_bytes",
          "warnings"
        ],
        "title": "AiPlanOut"
      },
      "AiRunIn": {
        "properties": {
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "device": {
            "type": "string",
            "enum": [
              "auto",
              "cpu"
            ],
            "title": "Device",
            "description": "`cpu` forces the CPU even with a GPU.",
            "default": "auto"
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "model_id"
        ],
        "title": "AiRunIn"
      },
      "AiRunStartOut": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobOut"
          },
          "plan": {
            "$ref": "#/components/schemas/AiPlanOut"
          }
        },
        "type": "object",
        "required": [
          "job",
          "plan"
        ],
        "title": "AiRunStartOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "State"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "state",
          "progress",
          "message",
          "params",
          "result",
          "error",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "JobOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}